requirements.txt
flask
tensorflow
keras
gunicorn
pillow